requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
PyYAML==3.10
Werkzeug==0.9.4
argparse==1.2.1
awscli==0.12.0
bcdoc==0.4.3
betamax==0.1.6
botocore==0.12.0
colorama==0.2.5
distribute==0.6.34
docutils==0.10
falcon==0.1.7
gevent==1.0
-e git+https://github.com/sigmavirus24/github3.py.git@17e20e805536e6b6f9a7802923b3b03524719056#egg=github3.py-dev
greenlet==0.4.1
ipython==1.1.0
itsdangerous==0.23
jira-python==0.16
jmespath==0.0.2
mock==1.0.1
mongoengine==0.8.3
oauthlib==0.6.0
ply==3.4
py==1.4.19
pymongo==2.5.2
pytest==2.3.5
python-dateutil==2.1
python-mimeparse==0.1.4
requests==2.1.0
requests-oauthlib==0.4.0
six==1.3.0
tlslite==0.4.6
uritemplate.py==0.2.0
wheel==0.21.0
wsgiref==0.1.2
